LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/location-setter-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: GET
FAIL: Timed out waiting for notifyDone to be called
